requirements.txt
pandas==1.5.3
python-dateutil==2.8.2
sqlite3==0.0.1
streamlit==1.15.2
plotly==5.11.0
